cargo clippy --fix  with cargo 1.67.0 (8ecd4f20a 2023-01-10)

+++ library/container/mpegts/src/pat.rs
@@ -61,7 +61,7 @@ impl PatMuxer {
                 .write_u16::<BigEndian>(ele.program_number)?;
             /*PID*/
             self.bytes_writer
-                .write_u16::<BigEndian>(0xE000 | (ele.pid as u16))?;
+                .write_u16::<BigEndian>(0xE000 | ele.pid)?;
         }
 
         /*crc32*/
